refactor code and add alembic for db migration

+++ routers/main_router.py
@@ -26,23 +26,27 @@ async def search_manga(site: MangaSiteEnum, search_keyword: str):
     return mangas
 
 
+def read_img_to_b64(file_path: str) -> str:
+    with open(file_path, 'rb') as f:
+        b = f.read()
+        return base64.b64encode(b).decode("utf-8")
+
+
 @router.get('/index/{site}/{manga_page}', response_model=Manga)
 async def get_index(site: MangaSiteEnum, manga_page: str, db: Session = Depends(get_db)):
     manga_site = get_manga_site(site)
     url = get_idx_page(site, manga_page)
     manga = catalog.get_manga(site, url)
-    need_update = manga.last_update + timedelta(days=3) < datetime.now()
+    need_update = manga.last_update is None or manga.last_update + timedelta(days=3) < datetime.now()
     if manga is None or not manga.idx_retrieved or need_update:
-        if need_update:
+        if need_update and manga.thum_img is not None:
             os.remove(manga.thum_img)
         manga = await manga_site.get_index_page(url)
         crud.create_chapters(db, manga)
         crud.update_manga_meta(db, manga)
 
     if manga.thum_img is not None:
-        with open(manga.thum_img, 'rb') as f:
-            b = f.read()
-            manga.thum_img = base64.b64encode(b).decode("utf-8")
+        manga.thum_img = read_img_to_b64(manga.thum_img)
 
     return manga
 
@@ -70,3 +74,9 @@ async def get_chapter(site: MangaSiteEnum, manga_page: str, idx: int, m_type_int
 @router.delete('/all')
 async def delete_all(db: Session = Depends(get_db)):
     return crud.delete_all(db)
+
+@router.post('/init')
+async def init_db(db: Session = Depends(get_db)):
+    for site in list(MangaSiteEnum):
+        site = crud.create_manga_site(db, site)
+    return True
